+++ ega-data-api-htsget/src/main/java/eu/elixir/ega/ebi/htsget/config/OAuth2ResourceConfig.java
@@ -16,20 +16,8 @@
 package eu.elixir.ega.ebi.htsget.config;
 
 
-import eu.elixir.ega.ebi.commons.config.CachingMultipleRemoteTokenService;
-import eu.elixir.ega.ebi.commons.config.CachingRemoteTokenService;
-import eu.elixir.ega.ebi.commons.config.MyAccessTokenConverter;
-import eu.elixir.ega.ebi.commons.config.MyUserAuthenticationConverter;
-import eu.elixir.ega.ebi.commons.shared.service.FileDatasetService;
-import eu.elixir.ega.ebi.commons.shared.service.Ga4ghService;
-import eu.elixir.ega.ebi.commons.shared.service.JWTService;
-import eu.elixir.ega.ebi.commons.shared.service.UserDetailsService;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.cache.annotation.EnableCaching;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Primary;
-import org.springframework.context.annotation.Profile;
 import org.springframework.core.annotation.Order;
 import org.springframework.http.HttpMethod;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
@@ -38,12 +26,7 @@ import org.springframework.security.oauth2.config.annotation.web.configuration.E
 import org.springframework.security.oauth2.config.annotation.web.configuration.ResourceServerConfigurerAdapter;
 import org.springframework.security.oauth2.provider.authentication.BearerTokenExtractor;
 import org.springframework.security.oauth2.provider.authentication.TokenExtractor;
-import org.springframework.security.oauth2.provider.token.AccessTokenConverter;
-import org.springframework.security.oauth2.provider.token.RemoteTokenServices;
 import org.springframework.security.web.authentication.preauth.AbstractPreAuthenticatedProcessingFilter;
-import org.springframework.web.cors.CorsConfiguration;
-import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
-import org.springframework.web.filter.CorsFilter;
 import org.springframework.web.filter.OncePerRequestFilter;
 
 import javax.servlet.FilterChain;
@@ -90,68 +73,5 @@ public class OAuth2ResourceConfig extends ResourceServerConfigurerAdapter {
                 .and()
                 .csrf().disable();
     }
-
-    // This is a bit of a Hack! MitreID doesn't return 'user_name' but 'user_id', The
-    // customized User Authentication Converter simply changes the field name for extraction
-    @Bean
-    public AccessTokenConverter accessTokenConverter(final JWTService jwtService,
-                                                     final Ga4ghService ga4ghService,
-                                                     final FileDatasetService fileDatasetService,
-                                                     final UserDetailsService userDetailsService) {
-        return new MyAccessTokenConverter(
-                jwtService,
-                ga4ghService,
-                fileDatasetService,
-                new MyUserAuthenticationConverter(),
-                userDetailsService
-        );
-    }
-
-    @Profile("enable-aai")
-    @Primary
-    @Bean
-    public RemoteTokenServices remoteTokenServices(HttpServletRequest request,
-                                                   //public RemoteTokenServices combinedTokenServices(HttpServletRequest request,
-                                                   final @Value("${auth.server.url}") String checkTokenUrl,
-                                                   final @Value("${auth.server.clientId}") String clientId,
-                                                   final @Value("${auth.server.clientsecret}") String clientSecret,
-                                                   final @Value("${auth.zuul.server.url}") String zuulCheckTokenUrl,
-                                                   final @Value("${auth.zuul.server.clientId}") String zuulClientId,
-                                                   final @Value("${auth.zuul.server.clientsecret}") String zuulClientSecret,
-                                                   final AccessTokenConverter accessTokenConverter) {
-
-        final CachingMultipleRemoteTokenService remoteTokenServices = new CachingMultipleRemoteTokenService();
-
-        // EGA AAI
-        CachingRemoteTokenService b = new CachingRemoteTokenService();
-        b.setCheckTokenEndpointUrl(checkTokenUrl);
-        b.setClientId(clientId);
-        b.setClientSecret(clientSecret);
-        b.setAccessTokenConverter(accessTokenConverter);
-        remoteTokenServices.addRemoteTokenService(b);
-
-        // ELIXIR AAI
-        CachingRemoteTokenService a = new CachingRemoteTokenService();
-        a.setCheckTokenEndpointUrl(zuulCheckTokenUrl);
-        a.setClientId(zuulClientId);
-        a.setClientSecret(zuulClientSecret);
-        remoteTokenServices.addRemoteTokenService(a);
-
-        return remoteTokenServices;
-    }
-
-    @Bean
-    @Order(0)
-    public CorsFilter corsFilter() {
-        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
-        CorsConfiguration config = new CorsConfiguration();
-        config.setAllowCredentials(true);
-        config.addAllowedOrigin("*");
-        config.addAllowedHeader("*");
-        config.addAllowedMethod("*");
-        source.registerCorsConfiguration("/**", config);
-        return new CorsFilter(source);
-    }
-
 }
